admin/traces/api.py
"""Traces API - Distributed tracing.


OpenTelemetry-compatible trace export.

- DevOps: Distributed tracing
- PhD Dev: Request flow analysis
- QA: Performance debugging
"""

from __future__ import annotations

import logging
from typing import Optional
from uuid import uuid4

from django.utils import timezone
from ninja import Router
from pydantic import BaseModel

from admin.common.auth import AuthBearer

router = Router(tags=["traces"])
logger = logging.getLogger(__name__)


# =============================================================================
# SCHEMAS
# =============================================================================


class Span(BaseModel):
    """Trace span."""

    span_id: str
    trace_id: str
    parent_span_id: Optional[str] = None
    operation_name: str
    service_name: str
    start_time: str
    end_time: str
    duration_ms: int
    status: str  # ok, error
    attributes: dict
    events: list[dict]


class Trace(BaseModel):
    """Distributed trace."""

    trace_id: str
    root_span_id: str
    service_count: int
    span_count: int
    duration_ms: int
    status: str
    start_time: str


# =============================================================================
# ENDPOINTS - Trace Query
# =============================================================================


@router.get(
    "",
    summary="List traces",
    auth=AuthBearer(),
)
async def list_traces(
    request,
    service: Optional[str] = None,
    operation: Optional[str] = None,
    status: Optional[str] = None,
    min_duration_ms: Optional[int] = None,
    limit: int = 50,
) -> dict:
    """List traces.

    DevOps: Trace search.
    """
    return {
        "traces": [],
        "total": 0,
    }


@router.get(
    "/{trace_id}",
    response=Trace,
    summary="Get trace",
    auth=AuthBearer(),
)
async def get_trace(request, trace_id: str) -> Trace:
    """Get trace details."""
    return Trace(
        trace_id=trace_id,
        root_span_id="span-1",
        service_count=3,
        span_count=10,
        duration_ms=500,
        status="ok",
        start_time=timezone.now().isoformat(),
    )


@router.get(
    "/{trace_id}/spans",
    summary="Get spans",
    auth=AuthBearer(),
)
async def get_trace_spans(
    request,
    trace_id: str,
) -> dict:
    """Get trace spans.

    PhD Dev: Request flow.
    """
    return {
        "trace_id": trace_id,
        "spans": [],
        "total": 0,
    }


# =============================================================================
# ENDPOINTS - Services
# =============================================================================


@router.get(
    "/services",
    summary="List services",
    auth=AuthBearer(),
)
async def list_services(request) -> dict:
    """List traced services.

    DevOps: Service discovery.
    """
    return {
        "services": [
            {"name": "soma-gateway", "span_count": 1000},
            {"name": "soma-agent", "span_count": 500},
            {"name": "soma-llm", "span_count": 200},
        ],
        "total": 3,
    }


@router.get(
    "/services/{service_name}/operations",
    summary="Get operations",
    auth=AuthBearer(),
)
async def get_service_operations(
    request,
    service_name: str,
) -> dict:
    """Get service operations."""
    return {
        "service": service_name,
        "operations": [
            {"name": "POST /completions/chat", "avg_duration_ms": 500},
            {"name": "GET /agents", "avg_duration_ms": 50},
        ],
        "total": 2,
    }


# =============================================================================
# ENDPOINTS - Stats
# =============================================================================


@router.get(
    "/stats",
    summary="Get trace stats",
    auth=AuthBearer(),
)
async def get_stats(
    request,
    period: str = "1h",
) -> dict:
    """Get tracing statistics.

    DevOps: Performance overview.
    """
    return {
        "period": period,
        "total_traces": 1000,
        "error_rate": 0.02,
        "avg_duration_ms": 150,
        "p50_duration_ms": 100,
        "p95_duration_ms": 300,
        "p99_duration_ms": 500,
    }


@router.get(
    "/stats/errors",
    summary="Error traces",
    auth=AuthBearer(),
)
async def error_traces(
    request,
    limit: int = 20,
) -> dict:
    """Get error traces.

    QA: Error debugging.
    """
    return {
        "errors": [],
        "total": 0,
    }


# =============================================================================
# ENDPOINTS - Export
# =============================================================================


@router.post(
    "/export",
    summary="Export traces",
    auth=AuthBearer(),
)
async def export_traces(
    request,
    start_time: str,
    end_time: str,
    service: Optional[str] = None,
) -> dict:
    """Export traces.

    DevOps: Trace export.
    """
    export_id = str(uuid4())

    return {
        "export_id": export_id,
        "status": "generating",
        "format": "jaeger",
    }
